Remove SaveGame call after setting a recovered kerbal's mission data
This SaveGame call was added back in the KSP 1.0 days, with a commit
comment that says it fixes a bug related to CrewQueueRoster._ExtDataSet
being cleared as part of loading different data from a save.  That
commit comment unfortunately doesn't explain what the bug *was*, but I
don't see any problems with the SaveGame call removed, so it's likely
something that was fixed another way by later changes in KSP and/or this
mod.  (As a guess, maybe the CrewRandRSettings scenario's OnSave didn't
get called after the recovery handler, so the reload after the scene
change didn't see the changes made by SetLastMissionData.  But nowadays,
OnSave does get called after the recovery handler.)

The SaveGame call doesn't hurt anything as-is, but it needs to be
removed before changing the recovery event hook from
OnVesselRecoveryRequested to onVesselRecovered, because saving the game
in an onVesselRecovered event handler makes the kerbals in the recovered
vessel go MIA.