Catch all kinds of vessel recovery, not just from flight
The OnVesselRecoveryRequested event fires when the player clicks the
"Recover" button above the altimeter in the flight scene, which is the
usual way to recover a vessel.  But it's also possible to recover
vessels from the tracking station, or even from the space center scene
if the vessel is nearby.  CrewRandR didn't catch those events, so the
crew would't be put on R&R after recovering a vessel in one of those
ways.

The onVesselRecovered event fires when a vessel recovery actually
occurs, regardless of how it was initiated.  This catches recoveries
from the tracking station and the space center (as well as from flight),
so the crew of the recovered vessel are always correctly put on R&R.