Combine rootExists and cleanedRoot flags into a single flag
The values of the rootExists and cleanedRoot flags weren't important
individually; all that actually mattered was the result of the
expression "rootExists && !cleanedRoot", which indicated that
CleanManifest() should be called.  It's simpler to just have a single
flag which indicates that condition.

This also removes the need for the onEditorPodDeleted and
onEditorRestart handlers.  They were actually already unnecessary, since
their purpose was just to set the rootExists and (indirectly via
Update()) cleanedRoot flags to false when the root part is deleted, but
the onEditorPodPicked and onEditorLoad handlers overwrote both of those
flags anyway when a new root was placed later.  It would've been weird
to leave rootExists true after deleting the root part, but now that
there's no rootExists flag, there's no need to clear it just to avoid
that weirdness.