Call CleanManifest() even when "automatically select crew" is disabled
This mod's settings page has a checkbox for whether to automatically
assign crew to a ship's root part, using an algorithm that's aware of
R&R and won't assign vacationing kerbals.  But with that setting turned
off, previously the game's default crew assignment would be used
instead, and that's *not* aware of R&R so vacationing kerbals would end
up being assigned.  To prevent that, it's important to always call
CleanManifest(), so that it can at least remove the default-assigned
crew even if it's not going to assign replacements.  (The result is that
ships in the VAB/SPH will initially have no crew at all, which seems
appropriate when the "automatically select crew" setting is turned off.)