Fix O(n^2) relabeling of kerbals in the Astronaut Complex
The loop for setting kerbals' "ready in..." labels in the Astronaut
Complex was a mix of two approaches: it iterated through all the
kerbals' UIList items, and for each, it iterated through all the
kerbals' CrewListItem objects.  That's two nested loops each iterating
over kerbals, so each kerbal would get relabeled many times (however
many kerbals are in the available list) when only once is needed.  The
UIList items weren't actually being used for anything, so I've removed
the outer loop and left just the inner one to iterate through the
kerbals' CrewListItem objects and set their labels once each.

Of the two approaches, iterating through the AstronautComplex's UIList
would be preferable over using FindObjectsOfType to find the kerbals,
but there seems to be no way to get a kerbal's CrewListItem from its
UIListItem.  (I'd expect the UIListItem's Data property to be a
reference to it, but those are all null.)  So it looks like
FindObjectsOfType is necessary.

(This was a pretty minor bug: the extra iterations didn't actually hurt
anything, and although inefficient, it wasn't a performance problem.  It
just produced some unnecessary logspam.)